ref: switch project to esm, decouple script for modules, add test simulating real user package run
