+++ .eslintrc.yml
@@ -1,12 +1,11 @@
 env:
   node: true
-  commonjs: true
   es2021: true
 extends:
   - eslint:recommended
 parserOptions:
   ecmaVersion: latest
-  sourceType: script
+  sourceType: module
 rules:
   curly: error
   default-case: error
